_site/tomat/requirements.txt
https://www.djangoproject.com/download/1.6b1/tarball/#egg=django
psycopg2==2.5
South==0.7.6
satchless==1.0
pytils==0.2.3
django-annoying==0.7.7
easy-thumbnails==1.2
django-mptt==0.5.5
python-memcached==1.53
Pillow==2.3.1
Fabric==1.6.1
requests
pytz
djorm-ext-pgfulltext==0.8
django-robokassa==1.2
Django-Select2==4.2.2
django-simple-captcha==0.4.3
django-bootstrap3-datetimepicker==2.2.3